Started to add in update function

+++ javascripts/delete.js
@@ -1,4 +1,4 @@
-const resultContainer = document.querySelector('.result-container');
+//const resultContainer = document.querySelector('.result-container');
 const deleteRestaurantBtn = document.querySelector('.deleteRestaurant');
 
 
@@ -13,11 +13,43 @@ function deleteRestaurant() {
       'Content-Type': 'application/json',
     },
     })
-    .then(response => response.json()); //not sure if I need this here
+    .then(response => response.json())
+    .then(data => {
+        let restaurants = data;
+        restaurants.forEach(function(restaurant) {
+            const div = document.createElement('div');
+            div.innerHTML = restaurant.name + restaurant.cuisine + restaurant.price + restaurant.visited + restaurant.favorite +  restaurant.priority;
+            // let input = document.createElement("input");
+            // input.setAttribute("type", "hidden");
+            // input.setAttribute("id", restaurant.id);
+            // div.appendChild(input);
+            let updateButton = document.createElement("button");
+            updateButton.innerHTML = "Update";
+            updateButton.setAttribute("class", "updateRestaurant");
+            updateButton.setAttribute("id", restaurant.id);
+            div.appendChild(updateButton);
+
+            let deleteButton = document.createElement("button");
+            deleteButton.innerHTML = "Delete";
+            deleteButton.setAttribute("class", "deleteRestaurant");
+            deleteButton.setAttribute("id", restaurant.id);
+            div.appendChild(deleteButton);
+
+            resultContainer.appendChild(div);
+        });
+    })
 }
 
-deleteRestaurantBtn.addEventListener('click', function() {
-    resultContainer.innerHTML = '';
-    deleteRestaurant();
-    alert("Restaurant was deleted");
-})
+if(deleteRestaurantBtn){
+    deleteRestaurantBtn.addEventListener('click',  function() {
+        resultContainer.innerHTML = '';
+        deleteRestaurant();
+        alert("Restaurant was deleted");
+    })
+}
+
+// deleteRestaurantBtn.addEventListener('click', function() {
+//     resultContainer.innerHTML = '';
+//     deleteRestaurant();
+//     alert("Restaurant was deleted");
+// })
